clean-up, translate to english and get raw metadata
