+++ window_config.json
@@ -1,0 +1,1 @@
+{"geometry": "1920x1009+-1+-8"}
